backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0
pydantic>=2.5.0
python-jose>=3.3.0
httpx>=0.25.0
rich>=13.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
jinja2>=3.1.0
